# 🚀 feat: Optimize loading performance and enhance loader implementation