--- a/apps/frontend/app/app/layout.tsx
+++ b/apps/frontend/app/app/layout.tsx
@@ -4,18 +4,35 @@ import { SidebarProvider } from "@/components/ui/sidebar";
 import "../globals.css";
 import { Toaster } from "@/components/ui/toaster";
 import AppNavbar from "@/app/app/navbar/app-navbar";
-import { useState, useEffect } from "react";
-import AppPage from "./page";
+import { useState, useEffect, lazy, Suspense } from "react";
 import Loader from "@/components/loader/Loader";
+import { PerformanceMonitor } from "@/components/performance/PerformanceMonitor";
 import { usePathname } from "next/navigation";
+
+// Lazy load the main app page for faster initial load
+const AppPage = lazy(() => import("./page"));
+
+// Loading skeleton for the main content
+const MainContentSkeleton = () => (
+  <div className="p-6 space-y-6 animate-pulse">
+    <div className="h-8 bg-gray-300 rounded w-1/3"></div>
+    <div className="h-4 bg-gray-300 rounded w-3/4"></div>
+    <div className="h-4 bg-gray-300 rounded w-1/2"></div>
+    <div className="h-32 bg-gray-300 rounded"></div>
+  </div>
+);
 
 export default function AppLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
-  const [mainContent, setMainContentRaw] = useState<React.ReactNode>(<AppPage />);
-  const [loading, setLoading] = useState(false);
+  const [mainContent, setMainContentRaw] = useState<React.ReactNode>(
+    <Suspense fallback={<MainContentSkeleton />}>
+      <AppPage />
+    </Suspense>
+  );
+  const [loading, setLoading] = useState(true);
   const [currentSection, setCurrentSection] = useState<string>("models"); // Default section
   const pathname = usePathname();
 
@@ -28,19 +45,32 @@ export default function AppLayout({
     setLoading(true);
     setMainContentRaw(content);
     setCurrentSection(sectionKey);
-    setTimeout(() => setLoading(false), 1000);
+    // Shorter transition for better UX
+    setTimeout(() => setLoading(false), 600);
   };
 
+  // Initial loading sequence
   useEffect(() => {
-    setLoading(true);
-    const timeout = setTimeout(() => setLoading(false), 800); // Adjust as needed
-    return () => clearTimeout(timeout);
+    const initialLoadTimer = setTimeout(() => {
+      setLoading(false);
+    }, 1200); // Show initial loader for app startup
+
+    return () => clearTimeout(initialLoadTimer);
+  }, []);
+
+  // Route change loading
+  useEffect(() => {
+    if (!loading) { // Only show route change loading if not in initial load
+      setLoading(true);
+      const timeout = setTimeout(() => setLoading(false), 400);
+      return () => clearTimeout(timeout);
+    }
   }, [pathname]);
 
   return (
     <html lang="en" className="dark">
       <body>
-        {loading && <Loader />}
+        {loading && <Loader message={currentSection === "models" ? "Loading Tatami..." : `Loading ${currentSection}...`} />}
         <AppNavbar />
         <SidebarProvider defaultOpen={false} className="overflow-hidden">
           <div className="relative flex w-full overflow-x-hidden">
@@ -49,8 +79,11 @@ export default function AppLayout({
               id="main-content"
               className="relative flex-1 h-full overflow-auto pt-5 transition-all duration-300 ease-in-out"
             >
-              {mainContent}
+              <Suspense fallback={<MainContentSkeleton />}>
+                {mainContent}
+              </Suspense>
               <Toaster />
+              <PerformanceMonitor />
             </div>
           </div>
         </SidebarProvider>