--- a/apps/frontend/app/app/page.tsx
+++ b/apps/frontend/app/app/page.tsx
@@ -1,9 +1,14 @@
 import { CodeDiagramSection } from "@/app/app/diagram/code-diagram-section";
+import { memo } from "react";
 
-export default function AppPage() {
+const AppPage = memo(() => {
   return (
-    <main className="w-full  h-full p-10">
+    <main className="w-full h-full p-10">
       <CodeDiagramSection />
     </main>
   );
-}
+});
+
+AppPage.displayName = "AppPage";
+
+export default AppPage;