… apps/frontend/components/navbar/Navbar.tsx
… apps/frontend/components/performance/PerformanceMonitor.tsx
… apps/frontend/components/ui/optimized-image.tsx
… apps/frontend/middleware.ts
… apps/frontend/next.config.ts
… docs/performance-optimizations.md