@diegoTech14 2025-07-05
Hello hello @kabugatti PR looks good, just some changes to do:

1. Ensure that you keep the current style because I noticed that you made some style changes in the code panel

2. Ensure that the first time that I enter to the app it loads fast