@kabugatti 2025-07-05
Hi @diegoTech14 ,

Thanks for the feedback

Lemme work on it.

FYI: The performance metric card only shows up in dev mode. I can disable it permanently or remove it, which is best?