@kabugatti 2025-07-05
## ✅ Updates Based on Feedback

Hi @diegoTech14,  Thank you for the review! I've addressed all the requested changes:

### 🎨 **Code Panel Styling - FIXED**
- **Issue**: Style changes were made to the code editor
- **Solution**: Reverted all styling changes to maintain the original `hc-black` theme
- **Files Updated**: 
  - `components/editor/DynamicEditor.tsx` - Added theme prop support
  - `app/app/diagram/code-diagram-section.tsx` - Explicitly passes `hc-black` theme
- **Result**: Code panel now maintains its original dark theme and all styling configurations

### ⚡ **First-Time Loading Speed - OPTIMIZED**
- **Issue**: Need to ensure fast first-time app loading
- **Solution**: Reduced loading timeouts across the board
- **Changes Made**:
  - Initial app load: `1200ms → 600ms` (50% faster)
  - Section transitions: `600ms → 300ms` (50% faster) 
  - Route changes: `400ms → 200ms` (50% faster)
  - Loader message cycling: `800ms → 600ms` (25% faster)
- **Files Updated**: `app/app/layout.tsx`, `components/loader/Loader.tsx`
- **Result**: Significantly faster first-time loading experience

### 📊 **Performance Metrics Card - REMOVED**
- **Issue**: Performance metrics card should not appear
- **Solution**: Completely disabled the performance monitor
- **Changes Made**:
  - Commented out `<PerformanceMonitor />` in layout
  - Removed the import to prevent any compilation issues
  - The entire component is now commented out as shown in the file
- **Result**: No performance metrics card visible anywhere in the app

### 🧪 **Testing Completed**
- ✅ App loads faster on first visit
- ✅ Code editor maintains original styling (`hc-black` theme)
- ✅ No performance metrics card appears
- ✅ All loader animations work smoothly with faster transitions
- ✅ Monaco Editor still loads asynchronously without blocking UI

All optimizations from the original PR remain intact while addressing the specific styling and performance concerns. The app now provides the enhanced loading experience without any visual regressions.