@kabugatti 2025-07-10
> Hello sir, could you do something about that topic? if not, don't worry I think the PR is nice.

Thanks for the feedback, it's something that will most likely require some time and iterations, not just a quick fix, but I'm willing to put more effort and time into it. You can go ahead and merge the current solution, and then we can work on a solution to the performance iteratively with time.